requirements.txt
flask
flask-caching
Flask_SQLAlchemy
flask_hcaptcha
redis
requests
anonymizeip
humanize
bbcode
Pillow
piexif
passlib
stopforumspam_api
flask-paginate
htmlmin
python-slugify
markdown2
psutil
bs4
maxminddb-geolite2
six
